+++ programs/clearing_house/src/lib.rs
@@ -387,12 +387,12 @@ pub mod clearing_house {
         optimal_borrow_rate: u32,
         max_borrow_rate: u32,
         oracle_source: OracleSource,
-        initial_asset_weight: u128,
-        maintenance_asset_weight: u128,
-        initial_liability_weight: u128,
-        maintenance_liability_weight: u128,
-        imf_factor: u128,
-        liquidation_fee: u128,
+        initial_asset_weight: u32,
+        maintenance_asset_weight: u32,
+        initial_liability_weight: u32,
+        maintenance_liability_weight: u32,
+        imf_factor: u32,
+        liquidation_fee: u32,
         active_status: bool,
     ) -> Result<()> {
         handle_initialize_spot_market(
@@ -431,7 +431,7 @@ pub mod clearing_house {
         oracle_source: OracleSource,
         margin_ratio_initial: u32,
         margin_ratio_maintenance: u32,
-        liquidation_fee: u128,
+        liquidation_fee: u32,
         active_status: bool,
         name: [u8; 32],
     ) -> Result<()> {
@@ -519,8 +519,8 @@ pub mod clearing_house {
 
     pub fn update_perp_liquidation_fee(
         ctx: Context<AdminUpdatePerpMarket>,
-        liquidator_fee: u128,
-        if_liquidation_fee: u128,
+        liquidator_fee: u32,
+        if_liquidation_fee: u32,
     ) -> Result<()> {
         handle_update_perp_liquidation_fee(ctx, liquidator_fee, if_liquidation_fee)
     }
@@ -534,8 +534,8 @@ pub mod clearing_house {
 
     pub fn update_spot_market_liquidation_fee(
         ctx: Context<AdminUpdateSpotMarket>,
-        liquidator_fee: u128,
-        if_liquidation_fee: u128,
+        liquidator_fee: u32,
+        if_liquidation_fee: u32,
     ) -> Result<()> {
         handle_update_spot_market_liquidation_fee(ctx, liquidator_fee, if_liquidation_fee)
     }
@@ -579,11 +579,11 @@ pub mod clearing_house {
 
     pub fn update_spot_market_margin_weights(
         ctx: Context<AdminUpdateSpotMarket>,
-        initial_asset_weight: u128,
-        maintenance_asset_weight: u128,
-        initial_liability_weight: u128,
-        maintenance_liability_weight: u128,
-        imf_factor: u128,
+        initial_asset_weight: u32,
+        maintenance_asset_weight: u32,
+        initial_liability_weight: u32,
+        maintenance_liability_weight: u32,
+        imf_factor: u32,
     ) -> Result<()> {
         handle_update_spot_market_margin_weights(
             ctx,
@@ -626,7 +626,7 @@ pub mod clearing_house {
 
     pub fn update_perp_market_imf_factor(
         ctx: Context<AdminUpdatePerpMarket>,
-        imf_factor: u128,
+        imf_factor: u32,
     ) -> Result<()> {
         handle_update_perp_market_imf_factor(ctx, imf_factor)
     }
