+++ programs/clearing_house/src/validation/order.rs
@@ -123,8 +123,8 @@ fn validate_limit_order(
             valid_oracle_price.ok_or(ErrorCode::InvalidOracle)?,
             slot,
             market.amm.order_tick_size,
-            market.margin_ratio_initial as u128,
-            market.margin_ratio_maintenance as u128,
+            market.margin_ratio_initial,
+            market.margin_ratio_maintenance,
         )?;
 
         if order_breaches_oracle_price_limits {
@@ -257,8 +257,8 @@ pub fn validate_spot_order(
     slot: u64,
     step_size: u64,
     tick_size: u64,
-    margin_ratio_initial: u128,
-    margin_ratio_maintenance: u128,
+    margin_ratio_initial: u32,
+    margin_ratio_maintenance: u32,
     min_order_size: u64,
 ) -> ClearingHouseResult {
     match order.order_type {
@@ -289,8 +289,8 @@ fn validate_spot_limit_order(
     step_size: u64,
     tick_size: u64,
     min_order_size: u64,
-    margin_ratio_initial: u128,
-    margin_ratio_maintenance: u128,
+    margin_ratio_initial: u32,
+    margin_ratio_maintenance: u32,
 ) -> ClearingHouseResult {
     validate_base_asset_amount(order, step_size, min_order_size, order.reduce_only)?;
